+++ migrator/service_provider_migration.go
@@ -1,4 +1,4 @@
-package client
+package migrator
 
 import (
 	"context"
